# contracts/openapi/corebank-service.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:18081",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/internal/v1/portfolio": {
      "get": {
        "tags": [
          "internal-corebank-controller"
        ],
        "operationId": "portfolio",
        "parameters": [
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/InternalPortfolioRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInternalPortfolioResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "internal-corebank-controller"
        ],
        "operationId": "provisionPortfolio",
        "parameters": [
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InternalPortfolioProvisioningRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInternalPortfolioProvisioningResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/orders": {
      "post": {
        "tags": [
          "internal-corebank-controller"
        ],
        "operationId": "createOrder",
        "parameters": [
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/InternalOrderCreateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInternalOrderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/accounts/{accountId}/status": {
      "get": {
        "tags": [
          "internal-corebank-controller"
        ],
        "operationId": "accountStatus",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/InternalAccountStatusRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInternalAccountStatusResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "internal-corebank-controller"
        ],
        "operationId": "transitionAccountStatus",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InternalAccountStatusTransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInternalAccountStatusTransitionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/v1/orders/{clOrdId}/requery": {
      "get": {
        "tags": [
          "internal-corebank-controller"
        ],
        "operationId": "requeryOrder",
        "parameters": [
          {
            "name": "clOrdId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$",
              "type": "string"
            }
          },
          {
            "name": "attemptCount",
            "in": "query",
            "description": "1-based requery attempt count from the scheduler. Unresolved results escalate when this reaches recovery.max-retry-count.",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "description": "1-based requery attempt count from the scheduler. Unresolved results escalate when this reaches recovery.max-retry-count.",
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInternalOrderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/accounts/{accountId}/summary": {
      "get": {
        "tags": [
          "internal-corebank-controller"
        ],
        "operationId": "accountSummary",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/InternalAccountSummaryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInternalAccountPositionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/accounts/{accountId}/positions": {
      "get": {
        "tags": [
          "internal-corebank-controller"
        ],
        "operationId": "accountPosition",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/InternalAccountPositionRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInternalAccountPositionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/accounts/{accountId}/positions/list": {
      "get": {
        "tags": [
          "internal-corebank-controller"
        ],
        "operationId": "accountPositions",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/InternalAccountPositionsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListInternalAccountPositionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/accounts/{accountId}/orders": {
      "get": {
        "tags": [
          "internal-corebank-controller"
        ],
        "operationId": "accountOrderHistory",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/InternalAccountOrderHistoryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInternalAccountOrderHistoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/accounts/default": {
      "get": {
        "tags": [
          "internal-corebank-controller"
        ],
        "operationId": "defaultAccountStatus",
        "parameters": [
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/InternalAccountStatusRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInternalAccountStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "InternalPortfolioProvisioningRequest": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "integer",
            "format": "int64"
          },
          "memberNo": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "userMessageKey": {
            "type": "string"
          },
          "operatorCode": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiResponseInternalPortfolioProvisioningResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/InternalPortfolioProvisioningResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InternalPortfolioProvisioningResponse": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "integer",
            "format": "int64"
          },
          "accountNumber": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "idempotent": {
            "type": "boolean"
          },
          "memberId": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InternalOrderCreateRequest": {
        "required": [
          "accountId",
          "clOrdId",
          "price",
          "quantity",
          "side",
          "symbol"
        ],
        "type": "object",
        "properties": {
          "accountId": {
            "type": "integer",
            "format": "int64"
          },
          "clOrdId": {
            "minLength": 1,
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$",
            "type": "string"
          },
          "symbol": {
            "minLength": 1,
            "type": "string"
          },
          "side": {
            "minLength": 1,
            "type": "string"
          },
          "quantity": {
            "minimum": 0.0001,
            "exclusiveMinimum": false,
            "type": "number"
          },
          "price": {
            "minimum": 0.0001,
            "exclusiveMinimum": false,
            "type": "number"
          }
        }
      },
      "ApiResponseInternalOrderResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/InternalOrderResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InternalOrderResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int64"
          },
          "clOrdId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "externalSyncStatus": {
            "type": "string"
          },
          "idempotent": {
            "type": "boolean"
          },
          "orderQuantity": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "retriable": {
            "type": "boolean"
          },
          "escalationRequired": {
            "type": "boolean"
          },
          "attemptCount": {
            "type": "integer",
            "format": "int32"
          },
          "maxRetryCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "InternalAccountStatusTransitionRequest": {
        "required": [
          "actor",
          "memberId",
          "reason",
          "status"
        ],
        "type": "object",
        "properties": {
          "memberId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "minLength": 1,
            "pattern": "ACTIVE|FROZEN|CLOSED",
            "type": "string"
          },
          "reason": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "actor": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "context": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ApiResponseInternalAccountStatusTransitionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/InternalAccountStatusTransitionResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InternalAccountStatusTransitionResponse": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "integer",
            "format": "int64"
          },
          "memberId": {
            "type": "integer",
            "format": "int64"
          },
          "previousStatus": {
            "type": "string"
          },
          "newStatus": {
            "type": "string"
          },
          "changed": {
            "type": "boolean"
          },
          "eventId": {
            "type": "integer",
            "format": "int64"
          },
          "reason": {
            "type": "string"
          },
          "actor": {
            "type": "string"
          },
          "context": {
            "type": "string"
          },
          "asOf": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InternalPortfolioRequest": {
        "required": [
          "accountId",
          "symbol"
        ],
        "type": "object",
        "properties": {
          "accountId": {
            "type": "integer",
            "format": "int64"
          },
          "symbol": {
            "minLength": 1,
            "type": "string"
          }
        }
      },
      "ApiResponseInternalPortfolioResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/InternalPortfolioResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InternalPortfolioResponse": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "integer",
            "format": "int64"
          },
          "accountNo": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "dailySellLimit": {
            "type": "number"
          },
          "todaySellQuantity": {
            "type": "number"
          }
        }
      },
      "InternalAccountSummaryRequest": {
        "required": [
          "memberId"
        ],
        "type": "object",
        "properties": {
          "memberId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ApiResponseInternalAccountPositionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/InternalAccountPositionResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "InternalAccountStatusRequest": {
        "required": [
          "memberId"
        ],
        "type": "object",
        "properties": {
          "memberId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ApiResponseInternalAccountStatusResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/InternalAccountStatusResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InternalAccountStatusResponse": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "integer",
            "format": "int64"
          },
          "memberId": {
            "type": "integer",
            "format": "int64"
          },
          "accountNumber": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "orderEligible": {
            "type": "boolean"
          },
          "denialCode": {
            "type": "string"
          },
          "asOf": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InternalAccountPositionRequest": {
        "required": [
          "memberId",
          "symbol"
        ],
        "type": "object",
        "properties": {
          "memberId": {
            "type": "integer",
            "format": "int64"
          },
          "symbol": {
            "minLength": 1,
            "type": "string"
          }
        }
      },
      "InternalAccountPositionsRequest": {
        "required": [
          "memberId"
        ],
        "type": "object",
        "properties": {
          "memberId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ApiResponseListInternalAccountPositionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InternalAccountPositionResponse"
            }
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InternalAccountOrderHistoryRequest": {
        "required": [
          "memberId"
        ],
        "type": "object",
        "properties": {
          "memberId": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "maximum": 100,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponseInternalAccountOrderHistoryResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/InternalAccountOrderHistoryResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InternalAccountOrderHistoryResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}